requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiolimiter==1.2.1
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
attrs==25.4.0
banks==2.4.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
boto3==1.42.68
botocore==1.42.68
build==1.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
chromadb==1.5.4
ci-info==0.4.0
click==8.3.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.5
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
ecdsa==0.19.1
etelemetry==0.3.1
fastapi==0.135.1
ffmpeg-python==0.2.0
filelock==3.25.1
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
future==1.0.0
google-auth==2.49.1
googleapis-common-protos==1.73.0
greenlet==3.3.2
griffe==2.0.0
griffecli==2.0.0
griffelib==2.0.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.6.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==1.2.12
langchain-anthropic==1.3.5
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.19
langchain-text-splitters==1.1.1
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.17
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.16
llama-index-cli==0.5.5
llama-index-core==0.14.17
llama-index-embeddings-openai==0.5.2
llama-index-embeddings-voyageai==0.6.0
llama-index-indices-managed-llama-cloud==0.10.0
llama-index-instrumentation==0.5.0
llama-index-llms-anthropic==0.11.0
llama-index-llms-openai==0.6.26
llama-index-readers-file==0.5.6
llama-index-readers-llama-parse==0.6.0
llama-index-vector-stores-chroma==0.5.5
llama-index-workflows==2.16.1
llama-parse==0.6.54
looseversion==1.3.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nibabel==5.4.2
nipype==1.11.0
nltk==3.9.3
numpy==2.4.3
oauthlib==3.3.1
onnxruntime==1.24.3
openai==2.28.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pathlib==1.0.1
pillow==12.1.1
platformdirs==4.9.4
propcache==0.4.1
protobuf==6.33.5
prov==2.1.1
puremagic==1.30
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.27.2
PyMySQL==1.1.2
pyparsing==3.3.2
pypdf==6.8.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
pytz==2026.1.post1
pyxnat==1.6.4
PyYAML==6.0.3
rdflib==7.6.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.52.1
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tinytag==2.2.1
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
traits==7.1.0
transformers==5.3.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
voyageai==0.3.7
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wrapt==1.17.3
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
